Adding things.

+++ libft/ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: sescolas <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/01/08 11:55:58 by sescolas          #+#    #+#             */
-/*   Updated: 2017/01/20 12:05:42 by sescolas         ###   ########.fr       */
+/*   Updated: 2017/08/09 16:13:34 by sescolas         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,10 +18,10 @@ static int	ft_iswhitespace(char c)
 			c == '\r' || c == '\f' || c == '\v');
 }
 
-int			ft_atoi(const char *str)
+ssize_t		ft_atoi(const char *str)
 {
-	int sum;
-	int sign;
+	ssize_t	sum;
+	int 	sign;
 
 	sign = 1;
 	sum = 0;
